+++ packages/eas-cli/__mocks__/ora.ts
@@ -3,5 +3,7 @@ export default jest.fn().mockReturnValue({
     fail: jest.fn(),
     succeed: jest.fn(),
   }),
+  stop: jest.fn(),
+  stopAndPersist: jest.fn(),
   succeed: jest.fn(),
 });