+++ packages/eas-cli/src/credentials/android/actions/SetupBuildCredentials.ts
@@ -1,11 +1,11 @@
 import nullthrows from 'nullthrows';
-import ora from 'ora';
 
 import {
   AndroidAppBuildCredentialsFragment,
   AndroidKeystoreFragment,
 } from '../../../graphql/generated';
 import Log from '../../../log';
+import { ora } from '../../../ora';
 import { Context } from '../../context';
 import { MissingCredentialsNonInteractiveError } from '../../errors';
 import { AppLookupParams } from '../api/GraphqlClient';